this might be controversial, but hear me out
we need some spacing between the logo and the menu button, but I don't 
want the logo to always have a margin on the left.

this way the spacing's only there if there's a button in in the markup, 
and it'll work for any adjacent sibling.

but I had to re order the markup a bit so that he sidemenu wasn't in 
between.