make the recirc tout a modifier on c-block-group
there was a lot of overlap, and I'd prefer if these classes were "design 
semantic" rather than content semantic.

that's the same reason why I'd like to change it from `--with-ad` to 
`--right-rail`. technically speaking, anything can go in that 300px 
column to the right, and I'd rather than have these concrete references 
floating around in the css.