make this heading semantically meaningless
article templates already have an `h1` in the article header organism,
and there should only bre one h1 on a page.

I made this update on the website, but I want pattern lab in sync. Plus
the letter spacing rule is inherited from h1, and needs to be expclitly
set on this element now.