use a mixin here
when this component is used in an app, I don't want other people to have 
to remember to attach a specific class to any links that end up in the 
title area, so let's target the element.

also, this is an example of using a mixin to compose a class.