some banner suggestions (#49)
* simplify title only state

the same UI can be achieved without using a modifier class. this'll make 
implementing it in user-land easier.

* make this a mixin

I'd like to start looking at using mixins across the design system, and 
using mixins to compose up some of the more concrete classes

* use a mixin here

when this component is used in an app, I don't want other people to have 
to remember to attach a specific class to any links that end up in the 
title area, so let's target the element.

also, this is an example of using a mixin to compose a class.

* include text-only title variation

* remove specialized `text-banner` molecule markup

* accept a block for `dek` area in block molecule

* create `urgent-block` variation on `block` molecule

refactors the `text-banner` markup into a variation on the existing 
`block` molecule. went with `urgent` because the text is red, which 
implies urgency.

* use urgent block molecule in text banner