Button animation (#61)
* Aligning the slide title and meat with the edges of the slide photo. (#60)

* Aligning the slide title and meat with the edges of the slide photo.

* need the equivalent of a flex-box clearfix here

the `o-end-label` was flowing into the same flex row as the final image 
at larger screen sizes. 

I made the icon a block element because once the width of the end label 
grew, the unwrapped text node flowed into the same line as the icon

this felt like a gallery-specific issue, which is why I put these rules 
in a concrete class

* Aligning the slide title and meat with the edges of the slide photo. (#60)

* Aligning the slide title and meat with the edges of the slide photo.

* need the equivalent of a flex-box clearfix here

the `o-end-label` was flowing into the same flex row as the final image 
at larger screen sizes. 

I made the icon a block element because once the width of the end label 
grew, the unwrapped text node flowed into the same line as the icon

this felt like a gallery-specific issue, which is why I put these rules 
in a concrete class

* Adding px-to-rem node module.

* Updating package.

* Adding print icon back.

* Updating animation for all email submisstion.

* Removng pxtorem and print icon.

* Updating footer copy.

* Updating urgent block timestamp.